# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/users-cadidature": {
            "get": {
                "summary": "Retrieve a list of all users",
                "operationId": "5e19560743d6290c740541cbb1825143",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of users"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users-acepted": {
            "post": {
                "summary": "Retrieve a list of all users accepted",
                "operationId": "1ca60390f45ed7756d3c390be055c491",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of users accepted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users-rejected": {
            "post": {
                "summary": "Retrieve a list of all users rejected",
                "operationId": "2ee57dae02990ede737114619cfaecbc",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of users rejected"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "get": {
                "summary": "Retrieve a one users ",
                "operationId": "37b77653b63a8b8ef93f709b6ab123ed",
                "responses": {
                    "200": {
                        "description": "Successfully a users selected"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "summary": "Update an existing user",
                "operationId": "1dad7375086fd778989b745ad58526bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "description": "User's first name",
                                        "type": "string"
                                    },
                                    "residence": {
                                        "description": "User's residence",
                                        "type": "string"
                                    },
                                    "profilePicture": {
                                        "description": "URL or base64-encoded image of the user's profile picture",
                                        "type": "string"
                                    },
                                    "levelOfStudy": {
                                        "description": "User's level of study",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/store": {
            "post": {
                "summary": "Create a new user",
                "operationId": "217b1c11b834a540f8d3ac0f7b24ab4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "description": "User's first name",
                                        "type": "string"
                                    },
                                    "residence": {
                                        "description": "User's residence",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password (hashed)",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User's email (unique)",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "profilePicture": {
                                        "description": "URL or base64-encoded image of the user's profile picture",
                                        "type": "string"
                                    },
                                    "levelOfStudy": {
                                        "description": "User's level of study",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/candidater/formation": {
            "post": {
                "summary": "Submit a candidature for a formation",
                "operationId": "46fed3212589a342f14289a0c6ddc0bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "formationId": {
                                        "description": "ID of the formation to apply for",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature submitted successfully"
                    },
                    "403": {
                        "description": "You have already applied to this formation"
                    },
                    "404": {
                        "description": "Formation not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/candidater/{formationId}": {
            "patch": {
                "summary": "Accept or reject a user's candidature for a formation",
                "operationId": "369c7572b7cd720de6fd856fcf1903ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "formationId": {
                                        "description": "ID of the formation",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "Status of the candidature (accept or reject)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature status updated successfully"
                    },
                    "404": {
                        "description": "Formation not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}